using uint8_t to represent bit plane related values in hierarchical 4d codec

+++ source/Lib/Part2/Common/TransformMode/Hierarchical4DCodec.h
@@ -64,8 +64,8 @@ protected:
 public:
 	std::array<ProbabilityModel, number_of_probability_models> probability_models;
     Block4D mSubbandLF;
-	int mSuperiorBitPlane = 30;
-	int mInferiorBitPlane = 0;
+	uint8_t mSuperiorBitPlane = 30;
+    uint8_t mInferiorBitPlane = 0;
     int mPGMScale;
     uint32_t mTransformLength_t, mTransformLength_s, mTransformLength_v, mTransformLength_u;
     //LightfieldDimension<uint32_t> transform_dimension;
